backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
pydantic>=2.7
pydantic-settings>=2.3
python-multipart>=0.0.9
httpx>=0.27
orjson>=3.10
python-docx>=1.1
lxml>=5.2
pytest>=8.0
